使用redux的combineReducers方法拆分管理组件对应的reducer
