+++ front-project/src/contain/contains.jsx
@@ -1,9 +1,12 @@
+/**
+ * 容器组件
+ */
 import {connect} from 'react-redux'
 
 import Header from '../components/header/header'
 import {headInputBlur, headInputFocus} from '../redux/actions'
 
 export default connect(
-    state => ({ focused: state }),
+    headInputState => ({ focused: headInputState }),
     { headInputBlur, headInputFocus }
 )(Header)
